refactor(event): add support to use uuid
